refactor(api): extract RoomLink dispatch handling
